Update CompoundHet.pl
Changed CAVA_EFFECT to CAVA_IMPACT as this annotation field changed in cava. Changed the gene symbol check to regex. Earlier, string was checked if it is ==0 and in perl all strings pass the condition check of ==0 and hence no annotation was created. Changed code to print compound het and effects properly.